// apps/www/src/interface.ts
export interface ICategory {
  id: string
  name: string
  country_code: string
  category_id: string
  slug: string
  country_info: ICountryInfo
}
// ...
export interface ITeam {
  gender: string
  slug: string
  id: string
  team_id: string
  name: string
  short_name: string
  abbreviation: string
  icon: string
  qualifier: string
  statistics: ITeamStatistics
  players: IPlayer[]
  form: string[]
}

export interface IPlayer {
  id: string
  name: string
  rank?: number
  team: ITeam
}

export interface ISportCategorySummary {
  id: string
  name: string
  category_id: string
  competitions: ILeague[]
  competitions_count: number
  latest_competition: ILeague
  slug: string
  country_info: ICountryInfo
}

export interface ISport {
  id: string
  name: string
  sport_id: string
  slug: string
}

export interface ISportEventContext {
  sport: ISport
  season: ISeason
  competition: ILeague
  category: ICategory
}

export interface ISportEvent {
  id: string
  sport_event_id?: string
  start_time: string
  start_time_formatted: string
  start_date_formatted: string
  competitors: ITeam[]
  sport_event_context: ISportEventContext
}

export type ISportEventStatusType = 'not_started' | 'closed' | 'live'

export type ISportEventMatchStatusType =
  | 'not_started'
  | 'ended'
  | '1st_half'
  | '2nd_half'
  | 'overtime'

export interface ISportEventStatus {
  clock: {
    played: string
  }
  start_time_formatted: string
  home_score: number
  away_score: number
  status: ISportEventStatusType
  match_status: ISportEventMatchStatusType
  winner_id?: string
  match_tie?: boolean
}

export interface ISportEventStatistics {
  totals: {
    competitors: ITeam[]
  }
}
// ...
export interface IGameSummary {
  sport_event: ISportEvent
  sport_event_status: ISportEventStatus
  statistics: ISportEventStatistics
  home_vs_away: ITeamVsTeam
  home_profile?: ITeamProfile
  away_profile?: ITeamProfile
}

export interface ITeamVsTeam {
  last_meetings: IGameSummary[]
  next_meetings: IGameSummary[]
}
// ...
export interface IVenue {
  name: string
  id: string
  venue_id: string
  capacity: number
  city_name: string
  country_name: string
  map_coordinates: string
  country_code: string
}

export interface IJersey {
  type: 'home'
}

export interface ITeamProfile {
  sport: ISport
  competitor: ITeam
  category: ICategory
  players: IPlayer[]
  manager?: IPlayer
  venue: IVenue
  summary_text?: string
  jerseys: IJersey[]
  total_players: number
  foreign_players: number
  home_players: number
}